# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Jewish prayer times calculation and publisher management",
        "title": "Zmanim Lab API",
        "termsOfService": "https://zmanim-lab.com/terms",
        "contact": {
            "name": "Zmanim Lab Support",
            "email": "support@zmanim-lab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/cities/{cityId}/publishers": {
            "get": {
                "description": "Returns all publishers that have coverage for the specified city (via city, region, country, or continent level)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Get publishers for city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City ID",
                        "name": "cityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publishers serving this city",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_jcom-dev_zmanim-lab_internal_models.PublishersForCityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/continents": {
            "get": {
                "description": "Returns all continents with their city counts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geography"
                ],
                "summary": "List continents",
                "responses": {
                    "200": {
                        "description": "List of continents",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/countries": {
            "get": {
                "description": "Returns all countries with their city counts, optionally filtered by continent",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geography"
                ],
                "summary": "List countries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by continent code or name",
                        "name": "continent",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of countries",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/regions": {
            "get": {
                "description": "Returns regions/states with their city counts, filtered by country code or search query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geography"
                ],
                "summary": "List regions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by ISO country code (required if search not provided)",
                        "name": "country_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query for region name (min 2 chars)",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of regions",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Either country_code or search required",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jcom-dev_zmanim-lab_internal_models.PublisherForCity": {
            "type": "object",
            "properties": {
                "coverage_level": {
                    "type": "string"
                },
                "match_type": {
                    "description": "exact_city, region_match, country_match",
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "publisher_id": {
                    "type": "string"
                },
                "publisher_name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_jcom-dev_zmanim-lab_internal_models.PublishersForCityResponse": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "string"
                },
                "publishers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_jcom-dev_zmanim-lab_internal_models.PublisherForCity"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_handlers.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/internal_handlers.ResponseMeta"
                }
            }
        },
...
        "internal_handlers.ResponseMeta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token from Clerk authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}